grid generator to own class
