+++ lib/maze_magic.rb
@@ -1,4 +1,6 @@
+require 'forwardable'
 require "maze_magic/version"
+require "maze_magic/grid"
 
 module MazeMagic
 end
